cookie_app/cloud_functions/ot_get_cookies/requirements.txt
functions-framework==3.*
requests==2.31.0
pandas==2.1.3
google-cloud-datastore==2.19.0